requirements/base.txt
django-filter==2.1.0
  - Django [required: >=1.11, installed: 2.1.7]
    - pytz [required: Any, installed: 2018.9]
djangorestframework==3.9.2
psycopg2==2.7.7
pytest-django==3.4.8
  - pytest [required: >=3.6, installed: 4.3.1]
    - atomicwrites [required: >=1.0, installed: 1.3.0]
    - attrs [required: >=17.4.0, installed: 19.1.0]
    - more-itertools [required: >=4.0.0, installed: 7.0.0]
    - pluggy [required: >=0.7, installed: 0.9.0]
    - py [required: >=1.5.0, installed: 1.8.0]
    - setuptools [required: Any, installed: 39.1.0]
    - six [required: >=1.10.0, installed: 1.12.0]
python-dateutil==2.8.0
  - six [required: >=1.5, installed: 1.12.0]
python-dotenv==0.10.1
requests==2.21.0
  - certifi [required: >=2017.4.17, installed: 2019.3.9]
  - chardet [required: >=3.0.2,<3.1.0, installed: 3.0.4]
  - idna [required: >=2.5,<2.9, installed: 2.8]
  - urllib3 [required: >=1.21.1,<1.25, installed: 1.24.1]